+++ src/js/pages/Dashboard.js
@@ -1,0 +1,13 @@
+import { Col, Row } from "react-bootstrap";
+import DashboardPresentational from "../components/presentationals/DashboardPresentational"
+
+const Home = () => (
+    <Row>
+        <Col>
+            <DashboardPresentational />
+        </Col>
+    </Row>
+);
+
+
+export default Home;